Login logout with session is completed

+++ app/admin/manage_admin.php
@@ -3,14 +3,17 @@
 ?>
   <!-- Content Wrapper. Contains page content -->
   <div class="content-wrapper">
-    <!-- Content Header (Page header) -->
+    <div class="container-fluid">
+    <h2>Create New Admin</h2>
+    <hr>
     <section class="content-header">
        <a href="create_admin.php" class="btn btn-primary">Create Admin</a><br>
       
     </section>
     
     <!-- Main content -->
-    <br><BR><div class="box-body">
+    
+    <br><BR><div class="box-body well">
     <table id="example1" class="table table-bordered table-striped">
                 <thead>
                 <tr>
@@ -32,7 +35,7 @@
                     <td><?php error_reporting(0); echo $row['contact'] ;?></td>
                     <td> <a href="<?php  echo $row['picture'];?>"><img src="<?php  echo $row['picture'];?>" class="img-circle" width="50px" height="50px"></a></td>
                     <td><a href="update_admin.php?admin_id=<?php echo $row['admin_id'];?>" class="label label-info">Update</a>
-                      <a href="deleteadmin.php?admin_id=<?php echo $row['admin_id'];?>" class="label label-danger">Delete</a></td>
+                      <a href="deleteadmin.php?admin_id=<?php echo $row['admin_id'];?>" class="label label-danger" onclick="return confirm('Are you sure to delete this admin');">Delete</a></td>
                   </tr>
                   <?php
                 }
@@ -49,7 +52,7 @@
                 </tfoot>
               </table>
               </div>
-
+          </div>
     <!-- /.content -->
   </div>
   <!-- /.content-wrapper -->
